requirements.txt
streamlit
numpy
pandas
altair
reportlab
pillow
openpyxl
xlsxwriter

# Production dependencies
python-dotenv==1.0.0
python-json-logger==2.0.7
psutil==5.9.6
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Optional: Redis support
redis==5.0.1

# Optional: PostgreSQL support
psycopg2-binary==2.9.9
